Refactor exicon-to-csv package to follow project conventions
This commit addresses all PR feedback items:

1. Split index.ts into separate files following "one export per file" convention
   - Created utils.ts for helper functions
   - Created fetch-exicon-data.ts, generate-exicon-csv.ts
   - Created exicon-to-csv-string.ts, exicon-to-csv-file.ts
   - Updated index.ts to re-export all functions

2. Removed side effects from modules
   - Created src/scripts/ directory for CLI execution
   - Moved execution logic to fetch-exicon-script.ts and get-unique-tags-script.ts
   - get-unique-tags.ts now only exports the function

3. Removed unused csv-parse dependency from package.json

4. Updated spec file to match actual implementation

5. Improved error handling in exicon-to-csv-file.ts
   - Added descriptive error messages with file paths
   - Maintains silent overwrite behavior (no breaking changes)

6. Added test coverage for special characters in tags

All 35 tests passing. Backward compatible - no breaking changes to API.
