+++ packages/exicon-to-csv/src/scripts/fetch-exicon-script.ts
@@ -1,4 +1,4 @@
-import { exiconToCsvFile } from "./index.js";
+import { exiconToCsvFile } from "../index.js";
 
 const filePath = await exiconToCsvFile();
 console.log(`CSV file written to: ${filePath}`);
