+++ packages/exicon-to-csv/src/scripts/get-unique-tags-script.ts
@@ -1,0 +1,4 @@
+import { getUniqueTags } from "../get-unique-tags.js";
+
+const tags = await getUniqueTags();
+console.log(tags);
